Communication between our API and Keycloak
This commit addresses the need for seamless communication between our API and Keycloak, the identity and access management solution.
The existing codebase underwent a significant refactoring to enhance integration and improve overall system security.
