+++ src/main/java/org/grnet/pidmr/endpoint/ProviderEndpoint.java
@@ -3,8 +3,6 @@ package org.grnet.pidmr.endpoint;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.grnet.pidmr.dto.InformativeResponse;
 import org.grnet.pidmr.dto.ProviderDto;
-import org.grnet.pidmr.dto.ProviderRequest;
-import org.grnet.pidmr.dto.UpdateProviderDto;
 import org.grnet.pidmr.dto.Validity;
 import org.grnet.pidmr.pagination.PageResource;
 import org.eclipse.microprofile.openapi.annotations.Operation;
@@ -14,25 +12,15 @@ import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
-import org.grnet.pidmr.repository.ProviderRepository;
 import org.grnet.pidmr.service.DatabaseProviderService;
-import org.grnet.pidmr.util.ServiceUriInfo;
-import org.grnet.pidmr.validator.constraints.NotFoundEntity;
 
 import javax.inject.Inject;
-import javax.validation.Valid;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import javax.ws.rs.DELETE;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
-import javax.ws.rs.PATCH;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
@@ -49,9 +37,6 @@ public class ProviderEndpoint {
 
     @Inject
     DatabaseProviderService providerService;
-
-    @ConfigProperty(name = "server.url")
-    String serverUrl;
 
     @Tag(name = "Provider")
     @Operation(
@@ -114,187 +99,6 @@ public class ProviderEndpoint {
         var validity = providerService.validation(pid, type);
 
         return Response.ok().entity(validity).build();
-    }
-
-    @Tag(name = "Provider")
-    @Operation(
-            summary = "Create a new Provider.",
-            description = "Endpoint for creating a new Provider.")
-    @APIResponse(
-            responseCode = "201",
-            description = "Created.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = ProviderDto.class)))
-    @APIResponse(
-            responseCode = "400",
-            description = "Invalid request payload.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "404",
-            description = "Not Found.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "409",
-            description = "Provider already exists.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "500",
-            description = "Internal Server Error.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @POST
-    @Produces(value = MediaType.APPLICATION_JSON)
-    public Response create(@Valid @NotNull(message = "The request body is empty.") ProviderRequest request, @Context UriInfo uriInfo) {
-
-        var response = providerService.create(request);
-
-        var serverInfo = new ServiceUriInfo(serverUrl.concat(uriInfo.getPath()));
-
-        return Response.created(serverInfo.getAbsolutePathBuilder().path(String.valueOf(response.id)).build()).entity(response).build();
-    }
-
-    @Tag(name = "Provider")
-    @Operation(
-            summary = "Get Provider by ID.",
-            description = "Endpoint for retrieving a Provider by its ID.")
-    @APIResponse(
-            responseCode = "200",
-            description = "The corresponding Provider.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = ProviderDto.class)))
-    @APIResponse(
-            responseCode = "404",
-            description = "Not Found.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "500",
-            description = "Internal Server Error.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @GET
-    @Path("/{id}")
-    @Produces(value = MediaType.APPLICATION_JSON)
-    public Response getProviderById(@Parameter(
-            description = "The ID of the Provider to retrieve.",
-            required = true,
-            example = "1",
-            schema = @Schema(type = SchemaType.NUMBER)) @PathParam("id")
-                                        @Valid @NotFoundEntity(repository = ProviderRepository.class, message = "There is no Provider with the following id:") Long id) {
-
-        var response = providerService.getProviderById(id);
-
-        return Response.ok().entity(response).build();
-    }
-
-    @Tag(name = "Provider")
-    @Operation(
-            summary = "Delete a Provider by ID.",
-            description = "Endpoint for deleting a Provider by its ID.")
-    @APIResponse(
-            responseCode = "200",
-            description = "The Provider has been successfully deleted.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "404",
-            description = "Not Found.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "500",
-            description = "Internal Server Error.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @DELETE
-    @Path("/{id}")
-    @Produces(value = MediaType.APPLICATION_JSON)
-    public Response deleteProviderById(@Parameter(
-            description = "The ID of the Provider to delete.",
-            required = true,
-            example = "1",
-            schema = @Schema(type = SchemaType.NUMBER)) @PathParam("id")
-                                    @Valid @NotFoundEntity(repository = ProviderRepository.class, message = "There is no Provider with the following id:") Long id) {
-
-        var deleted = providerService.deleteProviderById(id);
-
-        var response = new InformativeResponse();
-
-        if(deleted){
-
-            response.code = 200;
-            response.message = "The Provider has been successfully deleted.";
-            return Response.ok().entity(response).build();
-        } else {
-
-            response.code = 500;
-            response.message = "A problem occurred during the Provider deletion.";
-            return Response.serverError().entity(response).build();
-        }
-    }
-
-    @Tag(name = "Provider")
-    @Operation(
-            summary = "Update an existing Provider.",
-            description = "Endpoint for updating an existing Provider by ID. You can update a part or all attributes of Provider. The empty or null values are ignored.")
-    @APIResponse(
-            responseCode = "200",
-            description = "Updated.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = ProviderDto.class)))
-    @APIResponse(
-            responseCode = "400",
-            description = "Invalid request payload.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "404",
-            description = "Not Found.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "409",
-            description = "Provider already exists.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @APIResponse(
-            responseCode = "500",
-            description = "Internal Server Error.",
-            content = @Content(schema = @Schema(
-                    type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
-    @PATCH
-    @Path("/{id}")
-    @Produces(value = MediaType.APPLICATION_JSON)
-    public Response update(@Parameter(
-            description = "The ID of the Provider to update.",
-            required = true,
-            example = "1",
-            schema = @Schema(type = SchemaType.NUMBER)) @PathParam("id")
-                               @Valid @NotFoundEntity(repository = ProviderRepository.class, message = "There is no Provider with the following id:") Long id,
-                           @Valid @NotNull(message = "The request body is empty.") UpdateProviderDto request) {
-
-        var response = providerService.update(request, id);
-
-        return Response.ok().entity(response).build();
     }
 
     @Tag(name = "Provider")
