+++ src/main/resources/application.properties
@@ -1,17 +1,11 @@
 # This json file should encapsulate the list of Providers that the API supports
 list.providers.file = /etc/pidmr/conf.d/providers.conf
 
-%test.list.providers.file = providers.conf
-
 # This json file should encapsulate the list of Metaresolvers that the API supports
 list.metaresolvers.file = /etc/pidmr/conf.d/metaresolvers.conf
 
-%test.list.metaresolvers.file = metaresolvers.conf
-
 # This json file should encapsulate the list of Actions that the PID supports
 list.actions.file = /etc/pidmr/conf.d/actions.conf
-
-%test.list.actions.file = actions.conf
 
 
 # Specifies that each entry should be automatically removed from the cache once a fixed duration has elapsed after the entry's creation, the most recent replacement of its value, or its last read.
@@ -71,3 +65,29 @@ quarkus.flyway.baseline-version=1.0
 
 # The server URL that acts on behalf of the Cat Service
 server.url=${SERVER_URL:http://localhost:8080}
+
+%prod.quarkus.oidc.auth-server-url=https://login-devel.einfra.grnet.gr/auth/realms/einfra
+quarkus.oidc.client-id=backend-service
+quarkus.oidc.credentials.secret=secret
+quarkus.oidc.authorization-path=/protocol/openid-connect/auth
+quarkus.oidc.token-path=/protocol/openid-connect/token
+quarkus.oidc.discovery-enabled=false
+quarkus.oidc.introspection-path=/protocol/openid-connect/token/introspect
+%prod.quarkus.oidc.user-info-path=/protocol/openid-connect/userinfo
+%prod.quarkus.oidc.authentication.user-info-required=true
+
+# keycloak properties for feeding the oidc_client html template
+html.keycloak.server.url=http://localhost:58080
+html.keycloak.server.realm=quarkus
+html.keycloak.server.client.id=frontend-service
+html.keycloak.server.javascript.adapter=http://localhost:58080/js/keycloak.js
+html.cat.oidc.client.url=http://localhost:8080/oidc-client
+html.cat.api.documentation=http://localhost:8080/swagger-ui
+
+quarkus.keycloak.devservices.realm-path=quarkus-realm.json
+
+quarkus.keycloak.devservices.port=58080
+
+# Enable Policy Enforcement
+quarkus.keycloak.policy-enforcer.enable=true
+quarkus.keycloak.policy-enforcer.lazy-load-paths=false
