+++ src/main/resources/templates/client.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+		<script src="{keycloak_server_javascript_adapter}"></script>
+		<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
+		<meta charset="UTF-8">
+		<link rel="shortcut icon" href="#" />
+		<title>Keycloak Client</title>
+	</head>
+	<body>
+		<div class="row mb-3 mt-3 md-3 mx-3">
+			<div class="col align-self-start" id="token">
+			</div>
+			<div class="col align-self-center" id="copy_token">
+			</div>
+			<div class="col align-self-end" id="logout">
+			</div>
+		</div>
+
+		<script>
+
+		let initConfig = {
+              onLoad: 'login-required',
+              checkLoginIframe: true,
+              checkLoginIframeInterval: 1,
+              pkceMethod: 'S256',
+              scope: "voperson_id"
+        };
+
+        var keycloak = new Keycloak(
+            {
+                "url": "{keycloak_server_url}",
+                "realm": "{keycloak_server_realm}",
+                "clientId": "{keycloak_server_client_id}"
+            }
+        );
+
+        keycloak.init(initConfig).then(function(authenticated) {
+
+
+            var tokenDiv = document.getElementById('token');
+
+            var _text = document.createElement("input");
+            _text.id = "token_value";
+            _text.value = keycloak.token;
+            _text.className="form-control"
+
+            _text.for = "token_value";
+
+            tokenDiv.appendChild(_text);
+
+            var copyTokenDiv = document.getElementById('copy_token');
+
+            var _copy = document.createElement("button");
+            _copy.type="button";
+           _copy.data = "Copy Access Token";
+           _copy.innerHTML = 'Copy Access Token';
+           _copy.className="btn btn-primary";
+           _copy.onclick = function()
+            {
+                 var copyText = document.getElementById("token_value");
+                 copyText.select();
+                 copyText.setSelectionRange(0, 99999);
+                 navigator.clipboard.writeText(copyText.value);
+            }
+
+            copyTokenDiv.appendChild(_copy);
+
+            var logoutDiv = document.getElementById('logout');
+
+            var _logout = document.createElement("button");
+            _logout.type="button";
+           _logout.data = "Logout";
+           _logout.innerHTML = 'Logout';
+           _logout.className="btn btn-primary";
+           _logout.onclick = function()
+            {
+                 keycloak.logout();
+            }
+
+            logoutDiv.appendChild(_logout);
+        }).catch(function() {
+            console.log('failed to initialize');
+        });
+    </script>
+	</body>
+</html>
