+++ src/main/resources/templates/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <meta charset="UTF-8">
-        <title>pid-meta-resolver - 1.0.0-SNAPSHOT</title>
+        <title>PID Meta Resolver</title>
         <style>
             body {
                 margin: 0;
@@ -194,9 +194,9 @@
                 <div class="left-column">
                     <h1>This is the PID Metasolver API</h1>
                     <p>via the FAIRCORE4EOSC project.</p>
-                    <a href="https://apimr.devel.argo.grnet.gr/swagger-ui/" class="cta-button">Visit API Documentation</a>
+                    <a href="{cat_api_documentation}" class="cta-button">Visit API Documentation</a>
                     <a href="https://pidmr.devel.argo.grnet.gr/" class="cta-button">Visit devel UI</a>
-
+                    <a href="{cat_oidc_client_url}" class="cta-button">Obtain an Access Token</a>
                                     
                 </div>
                 <div class="right-column">
