+++ src/test/java/org/grnet/pidmr/KeycloakTest.java
@@ -1,0 +1,14 @@
+package org.grnet.pidmr;
+
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.keycloak.client.KeycloakTestClient;
+
+@QuarkusTest
+public class KeycloakTest {
+
+    KeycloakTestClient keycloakClient = new KeycloakTestClient();
+
+    protected String getAccessToken(String userName) {
+        return keycloakClient.getAccessToken(userName);
+    }
+}
