--- src/test/java/org/grnet/pidmr/MetaResolverTestProfile.java
@@ -1,11 +1,0 @@
-package org.grnet.pidmr;
-
-import io.quarkus.test.junit.QuarkusTestProfile;
-
-public class MetaResolverTestProfile implements QuarkusTestProfile {
-
-    @Override
-    public boolean disableApplicationLifecycleObservers() {
-        return true;
-    }
-}
