+++ src/test/java/org/grnet/pidmr/ProviderEndpointTest.java
@@ -1,21 +1,13 @@
 package org.grnet.pidmr;
 
 import org.grnet.pidmr.dto.InformativeResponse;
-import org.grnet.pidmr.dto.ProviderDto;
-import org.grnet.pidmr.dto.ProviderRequest;
-import org.grnet.pidmr.dto.UpdateProviderDto;
 import org.grnet.pidmr.dto.Validity;
 import org.grnet.pidmr.endpoint.ProviderEndpoint;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.TestProfile;
 import io.restassured.http.ContentType;
-import org.grnet.pidmr.service.DatabaseProviderService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-
-import javax.inject.Inject;
-import java.util.Set;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -23,13 +15,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @QuarkusTest
-@TestProfile(MetaResolverTestProfile.class)
 @TestHTTPEndpoint(ProviderEndpoint.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class ProviderEndpointTest {
-
-    @Inject
-    DatabaseProviderService providerService;
 
     @Test
     public void pidNotEmpty(){
@@ -517,340 +505,5 @@ public class ProviderEndpointTest {
                 .as(InformativeResponse.class);
 
         assertEquals("This type {not_supported} is not supported.", informativeResponse.message);
-    }
-
-    @Test
-    public void createProviderNotValidAction(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-not-valid";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("regexp");
-        request.actions = Set.of("not_valid_action");
-
-        var informativeResponse = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(404)
-                .extract()
-                .as(InformativeResponse.class);
-
-        assertEquals("There is an action that is not supported.", informativeResponse.message);
-    }
-
-    @Test
-    public void createProvider(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-create-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("regexp");
-        request.actions = Set.of("resource");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-        assertEquals("test-create-provider", provider.type);
-
-        providerService.deleteProviderById(provider.id);
-    }
-
-    @Test
-    public void createProviderExists(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-exist-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("regexp");
-        request.actions = Set.of("resource");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-
-        var informativeResponse = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(409)
-                .extract()
-                .as(InformativeResponse.class);
-
-        assertEquals("This Provider type {test-exist-provider} exists.", informativeResponse.message);
-
-        providerService.deleteProviderById(provider.id);
-    }
-
-    @Test
-    public void createProviderWithRegex(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-regex-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("rege(x(es)?|xps?)");
-        request.actions = Set.of("resource", "metadata");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-        var validity = given()
-                .contentType(ContentType.JSON)
-                .queryParam("type", provider.type)
-                .queryParam("pid", "regexps")
-                .get("/validate")
-                .then()
-                .assertThat()
-                .statusCode(200)
-                .extract()
-                .as(Validity.class);
-
-        assertTrue(validity.valid);
-
-        providerService.deleteProviderById(provider.id);
-    }
-
-    @Test
-    public void createProviderWithoutActions(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-regex-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("rege(x(es)?|xps?)");
-
-        var response = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(400)
-                .extract()
-                .as(InformativeResponse.class);
-
-
-        assertEquals("resolution_modes should have at least one entry.", response.message);
-    }
-
-    @Test
-    public void createProviderWithoutRegexes(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-regex-provider";
-        request.description = "Test Provider.";
-        request.actions = Set.of("resource", "metadata");
-
-        var response = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(400)
-                .extract()
-                .as(InformativeResponse.class);
-
-
-        assertEquals("regexes should have at least one entry.", response.message);
-    }
-
-    @Test
-    public void getProvider(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-get-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("rege(x(es)?|xps?)");
-        request.actions = Set.of("resource", "metadata");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-        var response = given()
-                .contentType(ContentType.JSON)
-                .get("/{id}", provider.id)
-                .then()
-                .assertThat()
-                .statusCode(200)
-                .extract()
-                .as(ProviderDto.class);
-
-        assertEquals(request.name, response.name);
-        assertEquals(request.type, response.type);
-
-        providerService.deleteProviderById(provider.id);
-    }
-
-    @Test
-    public void getProviderNotFound(){
-
-        var response = given()
-                .contentType(ContentType.JSON)
-                .get("/{id}", 1000L)
-                .then()
-                .assertThat()
-                .statusCode(404)
-                .extract()
-                .as(InformativeResponse.class);
-
-        assertEquals("There is no Provider with the following id: "+1000L, response.message);
-    }
-
-    @Test
-    public void deleteProvider(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-delete-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("rege(x(es)?|xps?)");
-        request.actions = Set.of("resource", "metadata");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-        var response = given()
-                .contentType(ContentType.JSON)
-                .delete("/{id}", provider.id)
-                .then()
-                .assertThat()
-                .statusCode(200)
-                .extract()
-                .as(InformativeResponse.class);
-
-        assertEquals("The Provider has been successfully deleted.", response.message);
-    }
-
-    @Test
-    public void updateBasicProviderInformation(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-change-basic-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("rege(x(es)?|xps?)");
-        request.actions = Set.of("resource");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-        var updateRequest = new UpdateProviderDto();
-        updateRequest.name = "Update Test Provider.";
-        updateRequest.type = "test-update-provider";
-        updateRequest.description = "Update Test Provider.";
-
-        var updatedProvider = given()
-                .body(updateRequest)
-                .contentType(ContentType.JSON)
-                .patch("/{id}", provider.id)
-                .then()
-                .assertThat()
-                .statusCode(200)
-                .extract()
-                .as(ProviderDto.class);
-
-        assertEquals(updateRequest.type, updatedProvider.type);
-        assertEquals(updateRequest.name, updatedProvider.name);
-        assertEquals(updateRequest.description, updatedProvider.description);
-        assertEquals(request.regexes.stream().findFirst().get(), updatedProvider.regexes.stream().findFirst().get());
-        assertEquals(request.actions.stream().findFirst().get(), updatedProvider.actions.stream().findFirst().get().mode);
-
-        providerService.deleteProviderById(provider.id);
-    }
-
-    @Test
-    public void updateProviderRegexesAndActions(){
-
-        var request = new ProviderRequest();
-        request.name = "Test Provider.";
-        request.type = "test-change-regex-action-provider";
-        request.description = "Test Provider.";
-        request.regexes = Set.of("rege(x(es)?|xps?)");
-        request.actions = Set.of("resource");
-
-        var provider = given()
-                .body(request)
-                .contentType(ContentType.JSON)
-                .post()
-                .then()
-                .assertThat()
-                .statusCode(201)
-                .extract()
-                .as(ProviderDto.class);
-
-        var updateRequest = new UpdateProviderDto();
-        updateRequest.regexes = Set.of("this is text");
-        updateRequest.actions = Set.of("metadata");
-
-        var updatedProvider = given()
-                .body(updateRequest)
-                .contentType(ContentType.JSON)
-                .patch("/{id}", provider.id)
-                .then()
-                .assertThat()
-                .statusCode(200)
-                .extract()
-                .as(ProviderDto.class);
-
-        assertEquals(request.type, updatedProvider.type);
-        assertEquals(request.name, updatedProvider.name);
-        assertEquals(request.description, updatedProvider.description);
-        assertEquals(updateRequest.regexes.stream().findFirst().get(), updatedProvider.regexes.stream().findFirst().get());
-        assertEquals(updateRequest.actions.stream().findFirst().get(), updatedProvider.actions.stream().findFirst().get().mode);
-
-        providerService.deleteProviderById(provider.id);
     }
 }
